+++ elastic-job-core/src/test/java/com/dangdang/ddframe/job/AllJobTests.java
@@ -24,6 +24,8 @@ import org.junit.runners.Suite.SuiteClasses;
 import com.dangdang.ddframe.job.api.JobExecutionMultipleShardingContextTest;
 import com.dangdang.ddframe.job.exception.JobConflictExceptionTest;
 import com.dangdang.ddframe.job.exception.TimeDiffIntolerableExceptionTest;
+import com.dangdang.ddframe.job.integrate.monitor.MonitorServiceDisableTest;
+import com.dangdang.ddframe.job.integrate.monitor.MonitorServiceEnableTest;
 import com.dangdang.ddframe.job.integrate.std.dataflow.sequence.OneOffSequenceDataFlowElasticJobTest;
 import com.dangdang.ddframe.job.integrate.std.dataflow.sequence.StreamingSequenceDataFlowElasticJobTest;
 import com.dangdang.ddframe.job.integrate.std.dataflow.throughput.OneOffThroughputDataFlowElasticJobTest;
@@ -48,8 +50,6 @@ import com.dangdang.ddframe.job.internal.execution.ExecutionServiceTest;
 import com.dangdang.ddframe.job.internal.failover.FailoverListenerManagerTest;
 import com.dangdang.ddframe.job.internal.failover.FailoverNodeTest;
 import com.dangdang.ddframe.job.internal.failover.FailoverServiceTest;
-import com.dangdang.ddframe.job.internal.monitor.MonitorServiceDisableTest;
-import com.dangdang.ddframe.job.internal.monitor.MonitorServiceEnableTest;
 import com.dangdang.ddframe.job.internal.offset.OffsetNodeTest;
 import com.dangdang.ddframe.job.internal.offset.OffsetServiceTest;
 import com.dangdang.ddframe.job.internal.server.JobOperationListenerManagerTest;
