+++ elastic-job-core/src/test/java/com/dangdang/ddframe/job/integrate/monitor/SocketUtils.java
@@ -1,0 +1,28 @@
+package com.dangdang.ddframe.job.integrate.monitor;
+
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
+import java.net.Socket;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class SocketUtils {
+    
+    public static String sendCommand(final String command, final int monitorPort) throws IOException {
+        try (
+                Socket socket = new Socket("127.0.0.1", monitorPort);
+                BufferedReader reader = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+                BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(socket.getOutputStream()))
+            ) {
+            writer.write(command);
+            writer.newLine();
+            writer.flush();
+            return reader.readLine();
+        }
+    }
+}
