--- elastic-job-core/src/test/java/com/dangdang/ddframe/job/internal/monitor/MonitorServiceDisableTest.java
@@ -1,36 +1,0 @@
-/**
- * Copyright 1999-2015 dangdang.com.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * </p>
- */
-
-package com.dangdang.ddframe.job.internal.monitor;
-
-import java.io.IOException;
-
-import org.junit.Test;
-
-public final class MonitorServiceDisableTest extends AbstractMonitorServiceTest {
-    
-    private static final int MONITOR_PORT = -1;
-    
-    public MonitorServiceDisableTest() {
-        super(MONITOR_PORT);
-    }
-    
-    @Test(expected = IOException.class)
-    public void assertMonitorWithDumpCommand() throws IOException {
-        sendCommand(MonitorService.DUMP_COMMAND, 9000);
-    }
-}
