+++ elastic-job-core/src/test/java/com/dangdang/ddframe/job/internal/offset/OffsetServiceTest.java
@@ -33,7 +33,7 @@ import org.mockito.MockitoAnnotations;
 import org.unitils.util.ReflectionUtils;
 
 import com.dangdang.ddframe.job.api.JobConfiguration;
-import com.dangdang.ddframe.job.internal.AbstractBaseJobTest.TestJob;
+import com.dangdang.ddframe.job.integrate.AbstractBaseStdJobTest.TestJob;
 import com.dangdang.ddframe.job.internal.storage.JobNodeStorage;
