+++ elastic-job-core/src/test/java/com/dangdang/ddframe/job/internal/statistics/ProcessCountJobTest.java
@@ -29,7 +29,7 @@ import org.quartz.JobExecutionException;
 import org.unitils.util.ReflectionUtils;
 
 import com.dangdang.ddframe.job.api.JobConfiguration;
-import com.dangdang.ddframe.job.internal.AbstractBaseJobTest.TestJob;
+import com.dangdang.ddframe.job.integrate.AbstractBaseStdJobTest.TestJob;
 import com.dangdang.ddframe.job.internal.server.ServerService;
 
 public final class ProcessCountJobTest {
